+++ Changelog.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2021-05-05
+
+### Added
+
+* Introduced alias `new Obs()` to short `new Observable()`
+* Added `Subscription` return value for `subscribe()` which can be removed directly
+
 ## [1.1.2] - 2021-05-05
 
 ### Changed